requirements.txt
Django==4.2.25
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
PyYAML==6.0.2
uritemplate==4.2.0
tzdata==2025.2
